+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 ### Changes ###
 
 - Added jquery.flot.composeImages.js to distribution
+- Change jquery.flot.legend.js to use SVG and add it to the distribution
 
 
 ## [1.0.0]